Move maps from schema test to separate file
